third commit, working repository

+++ src/main/java/com/deepak/spring_boot_4/__demo_project/models/Product.java
@@ -5,6 +5,8 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 
+import java.math.BigDecimal;
+
 @Entity
 public class Product {
 
@@ -14,8 +16,8 @@ public class Product {
     private String barCode;
     private String item;
     private String category;
-    private double price;
-    private double discount;
+    private BigDecimal price;
+    private BigDecimal discount;
     private boolean available;
 
 
@@ -51,19 +53,19 @@ public class Product {
         this.category = category;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 
-    public double getDiscount() {
+    public BigDecimal getDiscount() {
         return discount;
     }
 
-    public void setDiscount(double discount) {
+    public void setDiscount(BigDecimal discount) {
         this.discount = discount;
     }
